rvs coord update

+++ kaggle_runner/runners/coordinator.py
@@ -321,11 +321,15 @@ if [ x"${PHASE}" = x"dev" ]; then
   export PS4='[Remote]: Line ${LINENO}: '
 
   if [ "x${ENABLE_RVS}" = x1 ]; then
-    [ -z $(pgrep -f 'jupyter-notebook') ] && bash -x ./rvs.sh $SERVER $PORT 2>&1 ||
-    screen -d -m bash -c "{ echo [REMOTE]: rvs log below.; bash -x ./rvs.sh $SERVER $PORT 2>&1; } | $NC --send-only --no-shutdown -w 120s -i $(( 3600 * 2 ))s $SERVER $CHECK_PORT";
+    if [ -z $(pgrep -f 'jupyter-notebook') ]; then
+        bash ./rvs.sh $SERVER $PORT 2>&1 &
+    else
+        screen -d -m bash -c "{ echo [REMOTE]: rvs log below.; bash -x ./rvs.sh $SERVER $PORT 2>&1; } | $NC --send-only --no-shutdown -w 120s -i $(( 3600 * 2 ))s $SERVER $CHECK_PORT"
+    fi
   fi &
+  conda activate base
   make install_dep >/dev/null;
-  unbuffer make toxic 2>&1 | unbuffer -p tee -a lstm_log | ( [ $USE_AMQP -eq 0 ] && $NC --send-only -w 120s -i $(( 60 * 5 ))s $SERVER $CHECK_PORT || cat - )
+  unbuffer make toxic 2>&1 | unbuffer -p tee -a toxic_log | ( [ $USE_AMQP -eq 0 ] && unbuffer -p $NC --send-only -w 120s -i $(( 60 * 5 ))s $SERVER $CHECK_PORT || unbuffer -p cat - )
 fi
 
 if [ x"${PHASE}" != x"dev" ]; then
